ReorganizeModuleBackends: Update tests for removal of module_backend_config
https://github.com/caikit/caikit/issues/54
